Move promote button dropdown to separate component for easier reuse
